kamre-backend/requirements.txt
Flask~=2.1.3
python-dateutil~=2.8.2
Flask-Cors~=3.0.10
flask-mongoengine~=1.0.0
cryptography~=37.0.4
flask-pyjwt~=0.1.9
PyJWT~=2.4.0